# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Offix API",
    "description": "API REST del backend de Offix, red social de oficios (UTN FRSF).\n\n### Cómo autenticarse\n1. `POST /api/v1/auth/registro` para crear un usuario.\n2. `POST /api/v1/auth/login` para obtener un `access_token` (JWT).\n3. Enviarlo en cada request protegida como header `Authorization: Bearer <access_token>` (en Swagger UI: botón **Authorize**, arriba a la derecha).\n\n### Para el equipo de QA (Hoppscotch)\nEl spec completo también está disponible en `/openapi.json` y como archivo estático en `docs/openapi.json` del repo — se puede importar directo en Hoppscotch como colección.\n\n### Estado del proyecto\nSprint 1 — flujo de reseña (HU-01) con validación de enlace simulada, sin envío real de mail/WhatsApp. Ver `docs/plan-sprint-1.md` para el detalle de alcance y decisiones.",
    "contact": {
      "name": "Equipo Offix - UTN FRSF"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Entorno local (iniciar_backend.bat)"
    }
  ],
  "paths": {
...
    "/api/v1/oferentes": {
      "get": {
        "tags": [
          "Oferentes"
        ],
        "summary": "Buscar profesionales (público)",
        "description": "RF7/RF8 — Búsqueda y consulta pública de perfiles profesionales. No requiere login.",
        "operationId": "buscar_oferentes_api_v1_oferentes_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Palabra clave: nombre o apellido",
              "title": "Q"
            },
            "description": "Palabra clave: nombre o apellido"
          },
          {
            "name": "categoria_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por categoría/oficio",
              "title": "Categoria Id"
            },
            "description": "Filtrar por categoría/oficio"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OferenteOut"
                  },
                  "title": "Response Buscar Oferentes Api V1 Oferentes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Oferentes"
        ],
        "summary": "Crear el perfil profesional del usuario autenticado",
        "description": "RF2/HU-02 — Creación del perfil profesional del oferente. El\n`id_oferente` resultante es siempre igual al `id_usuario` del usuario\nautenticado (comparten clave primaria).",
        "operationId": "crear_perfil_oferente_api_v1_oferentes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OferenteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OferenteOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "El usuario autenticado no tiene rol Oferente"
          },
          "409": {
            "description": "El usuario ya tiene perfil, o el DNI/CUIT ya está registrado"
          },
          "422": {
            "description": "Campo obligatorio faltante o DNI/CUIT/teléfono con formato inválido"
          }
        }
      }
    },
    "/api/v1/oferentes/{oferente_id}": {
      "get": {
        "tags": [
          "Oferentes"
        ],
        "summary": "Consultar el perfil de un profesional (público)",
        "description": "RF7 — Consulta pública de perfil profesional. No requiere login.",
        "operationId": "obtener_oferente_api_v1_oferentes__oferente_id__get",
        "parameters": [
          {
            "name": "oferente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oferente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OferenteOut"
                }
              }
            }
          },
          "404": {
            "description": "Oferente no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Oferentes"
        ],
        "summary": "Editar el perfil profesional propio",
        "description": "RF2 — Edición del perfil profesional. Solo el dueño puede editarlo.",
        "operationId": "actualizar_perfil_oferente_api_v1_oferentes__oferente_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "oferente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oferente Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OferenteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OferenteOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "El usuario autenticado no es el dueño del perfil"
          },
          "404": {
            "description": "Oferente no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categorias": {
      "get": {
        "tags": [
          "Categorías"
        ],
        "summary": "Listar categorías/oficios (público)",
        "description": "RF5 — Navegación pública, sin login requerido.",
        "operationId": "listar_categorias_api_v1_categorias_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoriaOut"
                  },
                  "type": "array",
                  "title": "Response Listar Categorias Api V1 Categorias Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorías"
        ],
        "summary": "Crear categoría (solo Administrador)",
        "description": "RF18 — Administración de categorías de oficios.",
        "operationId": "crear_categoria_api_v1_categorias_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token"
          },
          "403": {
            "description": "Requiere rol Administrador"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/categorias/{categoria_id}": {
      "put": {
        "tags": [
          "Categorías"
        ],
        "summary": "Editar categoría (solo Administrador)",
        "description": "RF18.",
        "operationId": "actualizar_categoria_api_v1_categorias__categoria_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "categoria_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Categoria Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token"
          },
          "403": {
            "description": "Requiere rol Administrador"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categorías"
        ],
        "summary": "Eliminar categoría (solo Administrador)",
        "description": "RF18.",
        "operationId": "eliminar_categoria_api_v1_categorias__categoria_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "categoria_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Categoria Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Falta token"
          },
          "403": {
            "description": "Requiere rol Administrador"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoriaCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "CategoriaCreate",
        "example": {
          "descripcion": "Instalaciones y reparaciones eléctricas.",
          "nombre": "Electricista"
        }
      },
      "CategoriaOut": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "id_categoria": {
            "type": "integer",
            "title": "Id Categoria"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "id_categoria",
          "estado"
        ],
        "title": "CategoriaOut"
      },
      "CategoriaUpdate": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado"
          }
        },
        "type": "object",
        "title": "CategoriaUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OferenteCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "dni_cuit": {
            "type": "string",
            "title": "Dni Cuit"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "numero_matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Matricula"
          },
          "latitud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitud"
          },
          "longitud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitud"
          },
          "radio_cobertura_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radio Cobertura Km"
          },
          "hora_inicio_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Inicio Atencion"
          },
          "hora_fin_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Fin Atencion"
          },
          "disponible_emergencia": {
            "type": "boolean",
            "title": "Disponible Emergencia",
            "default": false
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "dni_cuit",
          "telefono"
        ],
        "title": "OferenteCreate",
        "example": {
          "apellido": "Pérez",
          "categoria_id": 1,
          "descripcion": "Electricista matriculado, más de 10 años de experiencia.",
          "disponible_emergencia": true,
          "dni_cuit": "20-30111222-3",
          "nombre": "Juan",
          "numero_matricula": "EL-12345",
          "telefono": "+54 3564 400111"
        }
      },
      "OferenteOut": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "dni_cuit": {
            "type": "string",
            "title": "Dni Cuit"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "numero_matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Matricula"
          },
          "latitud": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitud"
          },
          "longitud": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitud"
          },
          "radio_cobertura_km": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radio Cobertura Km"
          },
          "hora_inicio_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Inicio Atencion"
          },
          "hora_fin_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Fin Atencion"
          },
          "disponible_emergencia": {
            "type": "boolean",
            "title": "Disponible Emergencia",
            "default": false
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "id_oferente": {
            "type": "integer",
            "title": "Id Oferente"
          },
          "estado_verificacion": {
            "type": "string",
            "title": "Estado Verificacion"
          },
          "promedio_calificacion": {
            "type": "string",
            "title": "Promedio Calificacion"
          },
          "cantidad_rechazos_acumulados": {
            "type": "integer",
            "title": "Cantidad Rechazos Acumulados"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "dni_cuit",
          "telefono",
          "id_oferente",
          "estado_verificacion",
          "promedio_calificacion",
          "cantidad_rechazos_acumulados"
        ],
        "title": "OferenteOut"
      },
      "OferenteUpdate": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "apellido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apellido"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "numero_matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Matricula"
          },
          "latitud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitud"
          },
          "longitud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitud"
          },
          "radio_cobertura_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radio Cobertura Km"
          },
          "hora_inicio_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Inicio Atencion"
          },
          "hora_fin_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Fin Atencion"
          },
          "disponible_emergencia": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disponible Emergencia"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "title": "OferenteUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}